requirements.txt
asgiref==3.2.10
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.44.0
django-autoslug==1.9.8
django-cors-headers==3.7.0
django-material-admin==1.8.5
django-rest-knox==4.1.0
django-richtextfield==1.6.1
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
oauthlib==3.1.1
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2020.1
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.26.5
whitenoise==5.1.0
